pyproject.toml
[tool.poetry]
name = "czytacz"
version = "0.1.0"
description = ""
authors = ["ModZero <modzero@modzero.xyz>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.ruff]
select = ["E4", "E7", "E9", "F", "N", "D2", "D3", "D4"]

[[tool.mypy.overrides]]
# It seems next version will have types defined
module = "feedparser"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic_postgresql_enum"
ignore_missing_imports = true

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
psycopg = {extras = ["binary"], version = "^3.1.13"}
sqlalchemy = "^2.0.23"
argon2-cffi = "^23.1.0"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
feedparser = "^6.0.10"
typer = {extras = ["all"], version = "^0.9.0"}
celery = "^5.3.6"
celery-types = "^0.20.0"
tldextract = "^5.1.1"
alembic-postgresql-enum = "^0.4.0"

[tool.poetry.scripts]
czytacz = "czytacz.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
